astroalign+astrometry/requirements.txt
astroalign==2.3.1
astropy==4.2
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
idna==2.10
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
pyerfa==1.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.1
sep==1.1.1
six==1.15.0
tifffile==2021.3.17
urllib3==1.26.4
